generate_nodeid now generates random id's
Since the sqlite address-space is lazy loading,
the old approach with incremental identifiers is
very slow, as it requires loading the complete
address space in memory before the highest identifier
is found. Generating a random uid and checking if
it already exists in the address space is much faster,
especially for hunderds of nodes.